Move new PTT capture dialog code to separate file.
